requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0

pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
email-validator==2.2.0

pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
